Revert "EK/chore/clean code"
